Add dedicated auth endpoint config for virtual machines
An alternative to the previous submitted patch:
https://review.openstack.org/#/c/642812

Added a new method to find the public auth endpoint from the identity
provider in the service catalog, and assumes the version is v3 for ks.
I imagine I'll need to edit this to handle other identity providers?
Don't know enough about it!
Also could use some advice on multiline styling.
Changed from uri to url in this patch to match with auth_url in the
methods and not uri from the configs.

Change-Id: I5d8fc5977014b196c34f4a59a30a7525bc778359